requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-telegram-bot==13.15
pydantic==2.8.2
